Added support for iOS 13, back to UIKit lifecycle
Added back the SceneDelegate and AppDelegate due to a bug in xcode 12 (https://stackoverflow.com/questions/64114768/using-an-empty-lldb-target-which-can-cause-slow-memory-reads-from-remote-devices) related to the UITests. Since the SwiftUI lifecycle is not compatibile with xcode 11 I had to change the paradigm back to UIKit in order to make compatibile this App also with Xcode 11. Now the UITests run correct (at least with xcode 11.7)